#13: fix lint issues
