+++ frontend/src/components/auth/RequireAuth.tsx
@@ -1,6 +1,7 @@
 import type { ReactNode } from 'react';
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '../../context/AuthContext';
+
+import { useAuth } from '../../context/useAuth';
 
 export function RequireAuth({ children }: { children: ReactNode }) {
   const { currentUser, loading } = useAuth();
